requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
exceptiongroup==1.1.3
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
packaging==23.1
playwright==1.37.0
pluggy==1.3.0
pyee==9.0.4
pytest==7.4.0
pytest-base-url==2.0.0
pytest-playwright==0.4.2
python-slugify==8.0.1
requests==2.31.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.4